requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
chromadb
prometheus-fastapi-instrumentator
pyjwt 
passlib[bcrypt]
python-dotenv
python-multipart
httpx 
python-jose[cryptography]
redis
pytest-asyncio